Extract condition groups of CRS into separate files
